PersonalKnowledgeWorker/requirements.txt
python-dotenv
langchain
langchain-community
langchain-openai
langchain-chroma
langchain-huggingface
beautifulsoup4
httpx
gradio
chromadb

